+++ shardingsphere-operator/pkg/reconcile/computenode/deployment_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package deployment
+package computenode
 
 import (
 	"context"
